requirements.txt
fastapi
uvicorn[standard]
databases
aiomysql
sqlalchemy
python-dotenv
pydantic
